+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 10.0.0-SNAPSHOT - unreleased
 
+### ⚙️ Technical
+* Client Error messages are now saved and sent in bulk on a timer.  This allows Hoist to use a digest mail and
+  generally manage any storm of error reports.
+
 ## 9.4.0 - 2021-10-15
 
 ### 🎁 New Features